requirements.txt
Django
Celery
biopython
mysqlclient
suffix-trees
